tests: Disable Xwayland in Phoc tests
Xwayland cannot launch in certain environments, such as Distrobox
containers. While it's possible that this could be fixed with additional
configuration, there is no need to have Xwayland running for these
tests, so just disable it.

Also pull in Phoc from Debian experimental so that we get a new enough
version for this.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/218>